fix(codex-responses-api): resolve -D warnings lint failures
CI build blocked on two -D warnings surfaced by cargo build --release:
- src/proxy/responses_handler.rs:9 — unused `debug` import from tracing
- src/protocol/responses.rs:615 — unnecessary `i64 -> i64` cast
  (operand is already i64 from upstream .as_i64())
No behavior change; release build clean.

+++ src/protocol/responses.rs
@@ -612,7 +612,7 @@ pub(crate) fn response_from_chat(
         .unwrap_or(0);
     let completion_tokens = usage.and_then(|u| u.get("completion_tokens")).and_then(|v| v.as_i64()).unwrap_or(0);
 
-    let input_tokens = (prompt_tokens.saturating_sub(cached_tokens)).max(0) as i64;
+    let input_tokens = (prompt_tokens.saturating_sub(cached_tokens)).max(0);
     let total_tokens = input_tokens + completion_tokens;
 
     // ── Build response ──
